[TASK] Cleanup LoginController
Use functionality from OpenIdConnectService
