+++ Classes/Controller/LoginController.php
@@ -17,33 +17,22 @@ declare(strict_types=1);
 
 namespace Causal\Oidc\Controller;
 
-use Causal\Oidc\Service\OAuthService;
-use TYPO3\CMS\Core\Configuration\ExtensionConfiguration;
+use Causal\Oidc\Service\OpenIdConnectService;
 use TYPO3\CMS\Core\Http\PropagateResponseException;
 use TYPO3\CMS\Core\Http\RedirectResponse;
 use TYPO3\CMS\Core\Http\ServerRequest;
-use TYPO3\CMS\Core\Http\ServerRequestFactory;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Frontend\ContentObject\ContentObjectRenderer;
 
 class LoginController
 {
     /**
-     * Global oidc settings
-     *
-     * @var array
-     */
-    protected array $settings;
-
-    /**
      * TypoScript configuration of this plugin
-     *
-     * @var array
      */
     protected array $pluginConfiguration = [];
 
     /**
-     * @var ContentObjectRenderer|null will automatically be injected, if this controller is called as a plugin
+     * will automatically be injected, if this controller is called as a plugin
      */
     public ?ContentObjectRenderer $cObj = null;
 
@@ -51,7 +40,7 @@ class LoginController
 
     public function __construct()
     {
-        $this->settings = GeneralUtility::makeInstance(ExtensionConfiguration::class)->get('oidc') ?? [];
+        $this->request = $GLOBALS['TYPO3_REQUEST'];
     }
 
     public function setContentObjectRenderer(ContentObjectRenderer $cObj)
@@ -69,6 +58,7 @@ class LoginController
      *
      * @param string $_ ignored
      * @param array|null $pluginConfiguration
+     * @throws PropagateResponseException
      */
     public function login(string $_, ?array $pluginConfiguration)
     {
@@ -76,45 +66,25 @@ class LoginController
             $this->pluginConfiguration = $pluginConfiguration;
         }
 
-        $this->request = ServerRequestFactory::fromGlobals();
         $loginType = $this->request->getParsedBody()['logintype'] ?? $this->request->getQueryParams()['logintype'] ?? '';
         if ($loginType === 'login') {
-            // performRedirectAfterLogin stops flow by emitting a redirect
-            $this->performRedirectAfterLogin();
+            $redirectUrl = $this->determineRedirectUrl();
+            $this->redirect($redirectUrl);
         }
-        $this->performRedirectToLogin($pluginConfiguration['authorizationUrlOptions.']);
-    }
 
-    protected function performRedirectToLogin(array $authorizationUrlOptions = [])
-    {
-        /** @var OAuthService $service */
-        $service = GeneralUtility::makeInstance(OAuthService::class);
-        $service->setSettings($this->settings);
-
-        if (session_id() === '') {
-            session_start();
-        }
-        if ($this->settings['enableCodeVerifier']) {
-            $codeVerifier = $this->generateCodeVerifier();
-            $codeChallenge = $this->convertVerifierToChallenge($codeVerifier);
-            $authorizationUrlOptions = $this->addCodeChallengeToOptions($codeChallenge, $authorizationUrlOptions);
-            $_SESSION['oidc_code_verifier'] = $codeVerifier;
-        }
-        $authorizationUrl = $service->getAuthorizationUrl($authorizationUrlOptions);
-
-        $state = $service->getState();
-        $_SESSION['oidc_state'] = $state;
-        $_SESSION['oidc_login_url'] = GeneralUtility::getIndpEnv('REQUEST_URI');
-        $_SESSION['oidc_authorization_url'] = $authorizationUrl;
-        unset($_SESSION['oidc_redirect_url']); // The redirect will be handled by this plugin
-
+        $authorizationUrl = $this->determineAuthorizationUrl($pluginConfiguration['authorizationUrlOptions.']);
         $this->redirect($authorizationUrl);
     }
 
-    protected function performRedirectAfterLogin()
+    protected function determineAuthorizationUrl(array $authorizationUrlOptions): string
     {
-        $redirectUrl = $this->determineRedirectUrl();
-        $this->redirect($redirectUrl);
+        $oidcService = GeneralUtility::makeInstance(OpenIdConnectService::class);
+        $authorizationUrl = $oidcService->generateOpenidConnectUri($authorizationUrlOptions);
+
+        // The redirect will be handled by this plugin
+        unset($_SESSION['oidc_redirect_url']);
+
+        return $authorizationUrl;
     }
 
     protected function determineRedirectUrl()
@@ -125,8 +95,8 @@ class LoginController
         }
 
         if (isset($this->pluginConfiguration['defaultRedirectPid'])) {
-            $defaultRedirectPid = $this->pluginConfiguration['defaultRedirectPid'];
-            if ((int)$defaultRedirectPid > 0) {
+            $defaultRedirectPid = (int)$this->pluginConfiguration['defaultRedirectPid'];
+            if ($defaultRedirectPid > 0) {
                 return $this->cObj->typoLink_URL(['parameter' => $defaultRedirectPid]);
             }
         }
@@ -134,29 +104,11 @@ class LoginController
         return '/';
     }
 
+    /**
+     * @throws PropagateResponseException
+     */
     protected function redirect(string $redirectUrl): void
     {
         throw new PropagateResponseException(new RedirectResponse($redirectUrl));
-    }
-
-    protected function generateCodeVerifier(): string
-    {
-        return bin2hex(random_bytes(64));
-    }
-
-    protected function convertVerifierToChallenge($codeVerifier): string
-    {
-        return rtrim(strtr(base64_encode(hash('sha256', $codeVerifier, true)), '+/', '-_'), '=');
-    }
-
-    protected function addCodeChallengeToOptions($codeChallenge, array $options): array
-    {
-        return array_merge(
-            $options,
-            [
-                'code_challenge' => $codeChallenge,
-                'code_challenge_method' => 'S256',
-            ]
-        );
     }
 }
